requirements.txt
#统一按照最新版
selenium
python-dotenv
moviepy
requests
ffmpeg-python



aiofiles
annotated-types
anyio
browser-cookie3
certifi
click
colorama
fastapi
h11
httpcore
httpx==0.27.0
idna
importlib_resources
lz4
markdown-it-py
mdurl
numpy
pycryptodomex
pydantic
pydantic_core
pyfiglet
Pygments
pypng
pywebio
pywebio-battery
PyYAML
qrcode
rich
sniffio
starlette
tornado
typing_extensions
ua-parser
user-agents
uvicorn
websockets
gmssl
tenacity